enhancement: Order 제거 후 OrderDetail 을 Order 로 변경
